Analysis_opvset - Updated to handle higher numbers of indexes
Previously automatic data handlind assumed every trial was in a separate folder. ie numbers 1-6 ONLY could be suffix and would correspond to a letter.

Updated version will accept all of the numbers after the underscore, and introduced a max_channels value and using mod operator (%) to automatically parse the numbers to the correct number of channels. Useful for people who keep mutliple trials in the same folder